@@ -22,65 +22,65 @@
     "start": "ng serve",
     "build:dev": "ng build",
     "build:dev:watch": "npm run build:dev && ng build --watch",
-    "build:prod": "ng build --prod --aot --extract-css",
+    "build:prod": "ng build --configuration production --aot",
     "lint:sass": "sass-lint -c .sass-lint.yml Client/**/*.scss --verbose",
     "lint:ts": "tslint -c tslint.json Client/**/*.ts"
   },
   "dependencies": {
-    "@angular-devkit/schematics": "^11.0.4",
-    "@angular/animations": "11.2.14",
-    "@angular/common": "11.2.14",
-    "@angular/compiler": "11.2.14",
-    "@angular/core": "11.2.14",
-    "@angular/forms": "11.2.14",
-    "@angular/platform-browser": "11.2.14",
-    "@angular/platform-browser-dynamic": "11.2.14",
-    "@angular/platform-server": "11.2.14",
-    "@angular/router": "11.2.14",
+    "@angular-devkit/schematics": "^15.2.0",
+    "@angular/animations": "15.2.0",
+    "@angular/common": "15.2.0",
+    "@angular/compiler": "15.2.0",
+    "@angular/core": "15.2.0",
+    "@angular/forms": "15.2.0",
+    "@angular/platform-browser": "15.2.0",
+    "@angular/platform-browser-dynamic": "15.2.0",
+    "@angular/platform-server": "15.2.0",
+    "@angular/router": "15.2.0",
     "@microsoft/signalr": "3.0.1",
-    "@ng-bootstrap/ng-bootstrap": "^8.0.0",
-    "@popperjs/core": "2.0.0",
+    "@ng-bootstrap/ng-bootstrap": "14.0.1",
+    "@popperjs/core": "2.11.6",
     "acorn": "^6.4.1",
     "acorn-dynamic-import": "4.0.0",
-    "bootstrap": "4.4.1",
+    "bootstrap": "5.2.3",
     "core-js": "^3.14.0",
-    "file-loader": "2.0.0",
+    "file-loader": "^6.2.0",
     "font-awesome": "4.7.0",
     "is-svg": ">=4.2.2",
     "isomorphic-fetch": "3.0.0",
     "jquery": "3.5.0",
-    "ngx-toastr": "^13.2.0",
+    "ngx-toastr": "15.2.2",
     "normalize.css": "8.0.0",
     "popper.js": "1.16.1",
     "rxjs": "^6.5.2",
     "rxjs-compat": "^6.5.2",
     "ssri": ">=8.0.1",
     "tslib": "^2.0.0",
-    "typedoc": "0.22.11",
-    "webpack-dev-server": "4.7.4",
-    "zone.js": "~0.10.2"
+    "typedoc": "1.0.0-dev.4",
+    "webpack-dev-server": "^4.11.1",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.14",
-    "@angular/cli": "11.2.14",
-    "@angular/compiler-cli": "11.2.14",
-    "@angular/language-service": "11.2.14",
+    "@angular-devkit/build-angular": "^15.2.0",
+    "@angular/cli": "15.2.0",
+    "@angular/compiler-cli": "15.2.0",
+    "@angular/language-service": "15.2.0",
     "@types/core-js": "2.5.0",
     "@types/hammerjs": "2.0.35",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "@types/protractor": "4.0.0",
     "@types/selenium-webdriver": "3.0.10",
-    "codelyzer": "^6.0.0",
+    "codelyzer": "^6.0.2",
     "eslint": "8.10.0",
     "handlebars": "^4.7.7",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.16",
-    "karma-chrome-launcher": "~3.1.0",
+    "jasmine-core": "4.5.0",
+    "jasmine-spec-reporter": "7.0.0",
+    "karma": "6.4.1",
+    "karma-chrome-launcher": "^3.1.1",
     "karma-cli": "^2.0.0",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "karma-jasmine": "5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
     "lodash": "^4.17.21",
     "merge": "2.1.1",
     "npm-watch": "0.5.0",
@@ -90,9 +90,9 @@
     "ts-helpers": "1.1.2",
     "ts-node": "~7.0.1",
     "tslint": "~6.1.0",
-    "typedoc": "0.22.11",
-    "typescript": "4.0.8",
-    "url-loader": "1.1.1",
-    "webpack": "^4.42.1"
+    "typedoc": "1.0.0-dev.4",
+    "typescript": "4.9.5",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.76.0"
   }
 }
